feat(vm): add OP_EXTEND prefix to support 32-bit operands on demand
【动机】
当前所有操作数均限制为16位（最大值65535），但常量池索引、跳转偏移量和局部变量数量在实际程序中可能突破此限制。直接升级所有池索引为32位会导致指令体积膨胀，影响缓存效率。需要一种兼容现有16位指令格式的按需扩展机制。

【原理】
引入 `OP_EXTEND` 前缀指令，将特定操作数按需扩展为32位，同时保持大多数指令的紧凑性：

**编译端（Compiler）**：`emit16` 检测数值 > `0xFFFF` 时，回溯到当前主指令起始位置，在其前插入 `OP_EXTEND`（1字节）、操作数索引（1字节）和高16位数据（2字节），随后继续写入低16位数据。

**运行端（VM）**：维护 `extensionMap[4]` 缓存数组。遇到 `OP_EXTEND` 时读取索引和高16位存入缓存；遇到主指令时重置操作数计数器，`readOperand()` 宏读取低16位后利用 `std::exchange` 提取并清零对应缓存的高16位，按位或还原完整的32位操作数。

此设计实现向下兼容，且绝大多数指令保持16位紧凑格式，仅在必要时挂载扩展前缀，平衡了性能与可扩展性。

+++ src/vm/VM.h
@@ -35,7 +35,7 @@ namespace jc {
         static constexpr int MAX_STACK = 65536;
 
         std::vector<Value> globalValues;
-        std::unordered_map<std::string, uint16_t> globalNamesToSlots;
+        std::unordered_map<std::string, uint32_t> globalNamesToSlots;
         std::unordered_map<std::string, NativeCallable> nativeBuiltins;
         std::unordered_map<std::string, Value> builtinClosures;    // ★ 新增：内置函数闭包缓存
         std::unordered_set<std::string> constGlobals;              // ★ 新增：const 变量追踪
@@ -117,8 +117,8 @@ namespace jc {
             const std::vector<Value>& args);
 
         // ★ 类型检查冷路径：让繁重的字符串操作离开核心循环
-        [[noreturn]] void triggerParamTypeError(const Value& val, uint16_t typeIdx, uint16_t nameIdx);
-        [[noreturn]] void triggerReturnTypeError(const Value& val, uint16_t typeIdx);
+        [[noreturn]] void triggerParamTypeError(const Value& val, uint32_t typeIdx, uint32_t nameIdx);
+        [[noreturn]] void triggerReturnTypeError(const Value& val, uint32_t typeIdx);
 
         std::string getTypeName(const Value& val);
         bool checkValueType(const Value& val, const std::string& typeStr);
@@ -162,14 +162,14 @@ namespace jc {
         void execIndexSet(uint8_t dims);
         void execSliceGet(uint8_t dims);
         void execSliceSet(uint8_t dims);
-        void execBuildMatrix(uint16_t shapeIdx);
+        void execBuildMatrix(uint32_t shapeIdx);
         void execIn();
         Value execReturn(bool& shouldExit);
         void populateRefParams(CallFrame& newFrame, const CompiledFunction* fn);
-        void execInvoke(uint8_t argc, uint16_t icIdx, bool isTailCall = false);
-        void execSuperInvoke(uint16_t nameIdx, uint8_t argc, bool isTailCall = false);
-        void execAssertParamType(const Value& val, uint16_t typeIdx, uint16_t nameIdx);
-        void execAssertReturnType(const Value& val, uint16_t typeIdx);
+        void execInvoke(uint8_t argc, uint32_t icIdx, bool isTailCall = false);
+        void execSuperInvoke(uint32_t nameIdx, uint8_t argc, bool isTailCall = false);
+        void execAssertParamType(const Value& val, uint32_t typeIdx, uint32_t nameIdx);
+        void execAssertReturnType(const Value& val, uint32_t typeIdx);
 
     public:
         VM();
